+++ src/labkey/Experiment.ts
@@ -272,6 +272,11 @@ export interface LineageOptions extends ExperimentJSONConverterOptions, RequestC
     lsids?: string[]
     /** Include parents in the lineage response. Defaults to true. */
     parents?: boolean
+    /**
+     * Optional material lineage type to filter response -- either "AliquotationOnly" or "DerivationOnly".
+     * Defaults to include all.
+     */
+    materialRunType?: string
 }
 
 /**
@@ -304,6 +309,9 @@ export function lineage(options: LineageOptions): XMLHttpRequest {
     }
     if (options.cpasType) {
         params.cpasType = options.cpasType;
+    }
+    if (options.materialRunType) {
+        params.materialRunType = options.materialRunType;
     }
 
     return request({