+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## XX
+- Add materialRunType for Experiment.LineageOptions
+
 ## 1.1.9 - 2021-02-23
 - Add enumeration for Security.PermissionRoles
 - Correct controller name "specimens-api" -> "specimen-api"